# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/api/v1/supported-assets": {
            "get": {
                "description": "Public endpoint, Auth required in /admin/api/v1/supported-assets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supported Assets"
                ],
                "summary": "List supported (network, tokens) groupings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SupportedAssetsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/payments/ezpay/v1/supported-assets": {
            "get": {
                "description": "Public endpoint, Auth required in /admin/api/v1/supported-assets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supported Assets"
                ],
                "summary": "List supported (network, tokens) groupings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SupportedAssetsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/payments/ezpay/v1/supported-assets/records": {
            "get": {
                "description": "Public endpoint. Returns enabled chain_tokens, optionally filtered by network.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supported Assets"
                ],
                "summary": "List supported asset records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Network filter (e.g. tron, ethereum)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/mdb.ChainToken"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/payments/ezpay/v1/supported-assets/{id}": {
            "get": {
                "description": "Public endpoint. Returns one chain_token row by ID. Returns 404 if the token is disabled.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supported Assets"
                ],
                "summary": "Get a supported asset",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ChainToken ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/mdb.ChainToken"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/payments/gmpay/v1/supported-assets": {
            "get": {
                "description": "Public endpoint, Auth required in /admin/api/v1/supported-assets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supported Assets"
                ],
                "summary": "List supported (network, tokens) groupings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SupportedAssetsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/payments/gmpay/v1/supported-assets/records": {
            "get": {
                "description": "Public endpoint. Returns enabled chain_tokens, optionally filtered by network.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supported Assets"
                ],
                "summary": "List supported asset records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Network filter (e.g. tron, ethereum)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/mdb.ChainToken"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/payments/gmpay/v1/supported-assets/{id}": {
            "get": {
                "description": "Public endpoint. Returns one chain_token row by ID. Returns 404 if the token is disabled.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supported Assets"
                ],
                "summary": "Get a supported asset",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ChainToken ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/mdb.ChainToken"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "mdb.ChainToken": {
            "type": "object",
            "properties": {
                "contract_address": {
                    "type": "string",
                    "example": "TR7NHqjeKQxGTCi8q8ZY4pL8otSzgjLj6t"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-16 12:00:00"
                },
                "decimals": {
                    "type": "integer",
                    "example": 6
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "min_amount": {
                    "type": "number",
                    "example": 1
                },
                "network": {
                    "type": "string",
                    "example": "tron"
                },
                "symbol": {
                    "type": "string",
                    "example": "USDT"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-04-16 12:00:00"
                }
            }
        },
...
        "response.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123456"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "response.NetworkTokenSupport": {
            "type": "object",
            "properties": {
                "network": {
                    "type": "string",
                    "example": "tron"
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "USDT",
                        "USDC"
                    ]
                }
            }
        },
        "response.SupportedAssetsResponse": {
            "type": "object",
            "properties": {
                "supports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.NetworkTokenSupport"
                    }
                }
            }
        }
    }
}